+++ foxes/models/turbine_types/__init__.py
@@ -10,3 +10,4 @@ from .null_type import NullType
 from .wsrho2PCt_from_two import WsRho2PCtFromTwo
 from .wsti2PCt_from_two import WsTI2PCtFromTwo
 from .lookup import FromLookupTable
+from .TBL_file import TBLFile