+++ CHANGELOG.md
@@ -637,7 +637,10 @@ This major version introduces the concept of `Engines` which handle the chunking
 - Engines:
   - Default number of processes is now os.cpu_count(), or os.process_cpu_count() for python3.13
   - New engine: `RayEngine` (short name `ray`), runs parallel computations based on  the [Ray package](https://docs.ray.io/en/latest/)
+- Models:
+  - New turbine type `TBLFile`: Reads power, ct, D, H, P_rated from a *.tbl file
 - Bug fixes:
   - Bug fixed in `WSProduct`, causing zero wind speed at regions not touched by wakes
-
+  - Bugs fixes in `FarmLayoutOutput`, concerning the writing of the layout csv file
+  
 **Full Changelog**: [https://github.com/FraunhoferIWES/foxes/commits/v1.1](https://github.com/FraunhoferIWES/foxes/commits/v1.1)